+++ src/main/java/com/smart/smartparking/service/IRecordService.java
@@ -13,5 +13,6 @@ import com.smart.smartparking.entity.Record;
  */
 public interface IRecordService extends IService<Record> {
     Record getByPlateNumber(String plateNumber,int pid);
-    Integer selectCount(int pid);
+    Integer selectCount(Long pid);
+
 }
