+++ src/main/java/com/smart/smartparking/service/impl/ParkingServiceImpl.java
@@ -2,14 +2,10 @@ package com.smart.smartparking.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.smart.smartparking.common.Result;
 import com.smart.smartparking.entity.Parking;
 import com.smart.smartparking.mapper.ParkingMapper;
 import com.smart.smartparking.service.IParkingService;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Map;
 
 /**
  * <p>
@@ -39,10 +35,10 @@ public class ParkingServiceImpl extends ServiceImpl<ParkingMapper, Parking> impl
     }
 
     @Override
-    public Integer selectParkingPid(String name) {
+    public Long selectParkingPid(String name) {
         QueryWrapper<Parking> queryWrapper = new QueryWrapper<>();
         Parking parking =getOne(queryWrapper.eq("name",name));
-        int pid = parking.getPid();
+        Long pid = parking.getPid();
         return pid;
     }
